Clean Code and add comment
